+++ src/main/java/com/fernando9825/alcaldiasvrestapi/models/dao/IPuestoDao.java
@@ -3,5 +3,5 @@ package com.fernando9825.alcaldiasvrestapi.models.dao;
 import com.fernando9825.alcaldiasvrestapi.models.entity.Puesto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IPuestoDao extends JpaRepository<Puesto, Long> {
+public interface IPuestoDao extends JpaRepository<Puesto, Integer> {
 }
