+++ src/main/java/com/fernando9825/alcaldiasvrestapi/models/services/interfaces/ITarifaService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface ITarifaService {
 
-    List<Tarifa> getAllTarifasByInstitucion(Long institucionId);
+    List<Tarifa> getAllTarifasByInstitucion(short institucionId);
 }
